refactor : point/dto 폴더 밑으로 이동
