+++ src/main/java/com/kakaotech/team14backend/point/dto/UsePointByPostRequestDTO.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.outer.point.dto;
+package com.kakaotech.team14backend.point.dto;
 
 public record UsePointByPostRequestDTO(Long postId) {
