+++ src/main/java/com/kakaotech/team14backend/point/presentation/PointController.java
@@ -13,10 +13,10 @@ import com.kakaotech.team14backend.point.appliation.UsePointUsecase;
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import com.kakaotech.team14backend.inner.post.usecase.SetPostInstaCountUsecase;
-import com.kakaotech.team14backend.outer.point.dto.UsePointByPopularPostRequestDTO;
-import com.kakaotech.team14backend.outer.point.dto.UsePointByPopularPostResponseDTO;
-import com.kakaotech.team14backend.outer.point.dto.UsePointByPostRequestDTO;
-import com.kakaotech.team14backend.outer.point.dto.UsePointByPostResponseDTO;
+import com.kakaotech.team14backend.point.dto.UsePointByPopularPostRequestDTO;
+import com.kakaotech.team14backend.point.dto.UsePointByPopularPostResponseDTO;
+import com.kakaotech.team14backend.point.dto.UsePointByPostRequestDTO;
+import com.kakaotech.team14backend.point.dto.UsePointByPostResponseDTO;
 import com.kakaotech.team14backend.point.appliation.PointService;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
